More code edits.
